Add more OOP classes (A-MoreOOP)
